refactor: fetchPost 중복 제거
