+++ src/components/Post/content/PostList.tsx
@@ -92,7 +92,7 @@ function BoardList() {
     }
 
     fetchData();
-  });
+  }, []);
 
   if (loading) {
     return <div>Loading...</div>;
